+++ resources/lang/en/blogr.php
@@ -2,5 +2,9 @@
 
 // translations for Happytodev/Blogr
 return [
-    //
+    'users' => [
+        'navigation_label' => 'Users',
+        'model_label' => 'User',
+        'plural_model_label' => 'Users',
+    ],
 ];